refactor: move map config to separate module
